funds manager server: swap 0x with LiFi
