+++ funds-manager/funds-manager-server/src/handlers.rs
@@ -14,8 +14,7 @@ use funds_manager_api::hot_wallets::{
     TransferToVaultRequest, WithdrawToHotWalletRequest,
 };
 use funds_manager_api::quoters::{
-    DepositAddressResponse, ExecuteSwapRequest, ExecuteSwapResponse, GetExecutionQuoteRequest,
-    GetExecutionQuoteResponse, WithdrawFundsRequest,
+    DepositAddressResponse, ExecuteSwapRequest, ExecuteSwapResponse, WithdrawFundsRequest,
 };
 use itertools::Itertools;
 use serde_json::json;
@@ -145,21 +144,18 @@ pub(crate) async fn get_deposit_address_handler(
 
 /// Handler for getting an execution quote
 pub(crate) async fn get_execution_quote_handler(
-    req: GetExecutionQuoteRequest,
+    _body: Bytes, // no body
+    query_params: HashMap<String, String>,
     server: Arc<Server>,
 ) -> Result<Json, warp::Rejection> {
-    // Fetch the quoter hot wallet information
-    let vault = DepositWithdrawSource::Quoter.vault_name();
-    let hot_wallet = server.custody_client.get_hot_wallet_by_vault(vault).await?;
-    let wallet = server.custody_client.get_hot_wallet_private_key(&hot_wallet.address).await?;
+    // Forward the query parameters to the execution client
     let quote = server
         .execution_client
-        .get_quote(req.buy_token_address, req.sell_token_address, req.sell_amount, &wallet)
+        .get_quote(query_params)
         .await
         .map_err(|e| warp::reject::custom(ApiError::InternalError(e.to_string())))?;
 
-    let resp = GetExecutionQuoteResponse { quote };
-    Ok(warp::reply::json(&resp))
+    Ok(warp::reply::json(&quote))
 }
 
 /// Handler for executing a swap
